feat: move fetch update user to its component
